refactor(encode): promote Opus/FLAC encoders to public pkg/audio/encode
The Opus and FLAC encoders were the ONLY thing pkg/sendspin used from
internal/server (the source decoders and TUI are used solely by the server
CLI). Promoting them to a public package therefore drops pkg/sendspin's
internal/server dependency entirely.

- Move opus_encoder.go / flac_encoder.go (+ tests) verbatim from
  internal/server to pkg/audio/encode (package server -> encode); add doc.go.
- Repoint role_player.go, server.go, server_client.go: import
  pkg/audio/encode and reference encode.{Opus,FLAC}Encoder /
  encode.New{Opus,FLAC}Encoder.

Phase 2 GATE MET: go list -deps confirms pkg/ now imports ZERO internal/
packages. internal/server (sources + TUI) is used only by cmd/sendspin-server
now, so it can stay private to the future server-CLI repo.

Phase 2 of the three-repo split.
