+++ pkg/audio/encode/opus_encoder_test.go
@@ -1,6 +1,6 @@
 // ABOUTME: Tests for Opus audio encoder
 // ABOUTME: Tests encoder creation, encoding, and format handling
-package server
+package encode
 
 import (
 	"testing"
